@@ -8,8 +8,8 @@ export const CalendarIcon = () => (
     <rect
       fill="none"
 
-      stroke-linejoin="round"
-      stroke-width="32"
+      strokeLinejoin="round"
+      strokeWidth="32"
       x="48"
       y="80"
       width="416"
@@ -28,16 +28,16 @@ export const CalendarIcon = () => (
     <path
       fill="none"
      
-      stroke-linejoin="round"
-      stroke-width="32"
-      stroke-linecap="round"
+      strokeLinejoin="round"
+      strokeWidth="32"
+      strokeLinecap="round"
       d="M128 48v32M384 48v32"
     />
     <path
       fill="none"
      
-      stroke-linejoin="round"
-      stroke-width="32"
+      strokeLinejoin="round"
+      strokeWidth="32"
       d="M464 160H48"
     />
   </svg>
