@@ -8,9 +8,9 @@ export const FilterIcon = () => (
     <path
       fill="none"
       stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      stroke-width="32"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="32"
       d="M32 144h448M112 256h288M208 368h96"
     />
   </svg>
