@@ -8,14 +8,14 @@ export const NorwayIcon = () => (
   >
     <path d="M32 96H480V416H32V96Z" fill="#E92B2B" />
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M200 96H144V232H32L32 288H144V416H200V288L480 288V232L200 232V96Z"
       fill="#2A2FAF"
     />
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M152 96V232C152 236.418 148.418 240 144 240H32V224H136V96H152ZM192 232V96H208V224L480 224V240L200 240C195.582 240 192 236.418 192 232ZM32 280H144C148.418 280 152 283.582 152 288V416H136V296H32V280ZM194.343 282.343C195.843 280.843 197.878 280 200 280L480 280V296L208 296V416H192V288C192 285.878 192.843 283.843 194.343 282.343Z"
       fill="#FBFBFB"
     />
