@@ -4,11 +4,11 @@ export const SearchIcon = () => (
     width="20"
     height="20"
     viewBox="0 0 24 24"
-    stroke-width="1.5"
+    strokeWidth="1.5"
   
     fill="none"
-    stroke-linecap="round"
-    stroke-linejoin="round"
+    strokeLinecap="round"
+    strokeLinejoin="round"
   >
     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
     <path d="M10 10m-7 0a7 7 0 1 0 14 0a7 7 0 1 0 -14 0" />
