@@ -4,10 +4,10 @@ export const SortIcon = () => (
     width="20"
     height="20"
     viewBox="0 0 24 24"
-    stroke-width="1.5"
+    strokeWidth="1.5"
     fill="none"
-    stroke-linecap="round"
-    stroke-linejoin="round"
+    strokeLinecap="round"
+    strokeLinejoin="round"
   >
     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
     <path d="M3 9l4 -4l4 4m-4 -4v14" />
