@@ -1,23 +1,33 @@
 import React from "react";
 import { Table, flexRender } from "@tanstack/react-table";
-import { Customer } from "./types";
+
+import { Loader } from "../../common/Loader";
+import { Customer, CustomerColumns } from "./useCustomers";
 
 export interface CustomersTableProps {
-  table: Table<Customer>;
+  table: Table<CustomerColumns>;
+  loading?: boolean;
 }
 
 const columnWidths = {
-  col0: '7%',
-  col1: '14%', 
-  col2: '14%',
-  col3: '17%',
-  col4: '15%',
-  col5: '19%',
-  col6: '15%',
-
+  col0: "7%",
+  col1: "14%",
+  col2: "14%",
+  col3: "17%",
+  col4: "15%",
+  col5: "19%",
+  col6: "15%",
 };
 
-export const CustomersTable = ({ table }: CustomersTableProps) => {
+export const CustomersTable = ({ table, loading }: CustomersTableProps) => {
+  if (loading) {
+    return (
+      <div className="w-full min-h-[50vh] flex justify-center items-center">
+        <Loader />
+      </div>
+    );
+  }
+
   return (
     <table className="w-full mt-8 min-w-[60rem]">
       <thead>
@@ -33,12 +43,13 @@ export const CustomersTable = ({ table }: CustomersTableProps) => {
                     : "text-secondaryText dark:text-secondaryTextDark font-normal text-left text-base pl-3 2xl:pl-5 py-3 border cursor-pointer select-none  bg-inputBg dark:bg-inputBgDark border-inputBorder dark:border-inputBorderDark"
                 }
                 onClick={header.column.getToggleSortingHandler()}
-                style={{ 
+                style={{
                   width: columnWidths[header.id as keyof typeof columnWidths],
-                   maxWidth: columnWidths[header.id as keyof typeof columnWidths],
-                   minWidth: columnWidths[header.id as keyof typeof columnWidths] 
-                  }}
-
+                  maxWidth:
+                    columnWidths[header.id as keyof typeof columnWidths],
+                  minWidth:
+                    columnWidths[header.id as keyof typeof columnWidths],
+                }}
               >
                 {header.isPlaceholder
                   ? null
