@@ -1,24 +1,23 @@
-import { useMemo } from "react";
+import { useEffect, useMemo, useState } from "react";
 import {
   createColumnHelper,
   getCoreRowModel,
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import { Customer } from "./types";
-import { customersData } from "./CustomersData";
+
 import { useTable } from "../../../hooks/useTable";
 import Image from "next/image";
 
-const columnHelper = createColumnHelper<Customer>();
+const columnHelper = createColumnHelper<CustomerColumns>();
 
 const customerColumns = [
   columnHelper.accessor("col0", {
     header: () => "Photo",
     enableSorting: false,
-    cell: ({ row }: { row: { original: Customer } }) => (
+    cell: ({ row }: { row: { original: CustomerColumns } }) => (
       <Image
-        src={row.original.col0.src}
+        src={row.original.col0}
         alt="User Profile"
         width={50}
         height={50}
@@ -42,7 +41,7 @@ const customerColumns = [
   }),
   columnHelper.accessor("col6", {
     header: () => "Total Buys",
-    cell: ({ row }: { row: { original: Customer } }) =>
+    cell: ({ row }: { row: { original: CustomerColumns } }) =>
       row.original.col6.toString(),
   }),
 ];
@@ -50,8 +49,26 @@ const customerColumns = [
 export type CustomerFilters = {
   country?: string;
 };
+export interface CustomerColumns {
+  col0: string; // Avatar
+  col1: string; // First Name
+  col2: string; // Last Name
+  col3: string; // City
+  col4: string; // Country
+  col5: string; // Phone
+  col6: number; // Total Buys
+}
+export interface Customer {
+  photo: string;
+  firstName: string;
+  lastName: string;
+  city: string;
+  country: string;
+  phone: string;
+  totalBuys: number;
+}
 
-export const useCustomers = () => {
+export const useCustomers = (customers: Customer[]) => {
   const {
     sorting,
     setSorting,
@@ -68,12 +85,28 @@ export const useCustomers = () => {
     clearFilter,
     setFilter,
   } = useTable<CustomerFilters>({});
+  const [customersData, setCustomersData] = useState<Customer[]>();
+
+  useEffect(() => {
+    if (customers) {
+      const mappedData = customers.map((customer) => ({
+        photo: customer.photo,
+        firstName: customer.firstName,
+        lastName: customer.lastName,
+        city: customer.city,
+        country: customer.country,
+        phone: customer.phone,
+        totalBuys: customer.totalBuys,
+      }));
+      setCustomersData(mappedData);
+    }
+  }, [customers]);
 
   const filteredData = useMemo(() => {
     let data = customersData;
     setCurrentPage(0);
     if (searchQuery) {
-      data = data.filter((customer) =>
+      data = data?.filter((customer) =>
         Object.values(customer).some(
           (value) =>
             value &&
@@ -83,19 +116,27 @@ export const useCustomers = () => {
     }
 
     if (filters.country) {
-      data = data.filter((customer) => customer.col4 === filters.country);
+      data = data?.filter((customer) => customer.country === filters.country);
     }
 
     return data;
-  }, [searchQuery, filters]);
+  }, [searchQuery, filters, customersData]);
 
   const sortedAndFilteredData = useMemo(() => {
-    let data = [...filteredData];
+    let data = (filteredData ?? []).map((customer) => ({
+      col0: customer.photo,
+      col1: customer.firstName,
+      col2: customer.lastName,
+      col3: customer.city,
+      col4: customer.country,
+      col5: customer.phone,
+      col6: customer.totalBuys,
+    }));
 
     sorting.forEach(({ id, desc }) => {
       data.sort((a, b) => {
-        const aValue = a[id as keyof Customer];
-        const bValue = b[id as keyof Customer];
+        const aValue = a[id as keyof CustomerColumns];
+        const bValue = b[id as keyof CustomerColumns];
 
         if (typeof aValue === "number" && typeof bValue === "number") {
           return desc ? bValue - aValue : aValue - bValue;
@@ -115,9 +156,9 @@ export const useCustomers = () => {
     return sortedAndFilteredData.slice(startIdx, endIdx);
   }, [sortedAndFilteredData, currentPage, itemsPerPage]);
 
-  const totalPages = Math.ceil(filteredData.length / itemsPerPage);
+  const totalPages = Math.ceil((filteredData?.length ?? 0) / itemsPerPage);
 
-  const table = useReactTable<Customer>({
+  const table = useReactTable({
     columns: customerColumns,
     data: paginatedData,
     getCoreRowModel: getCoreRowModel(),
@@ -143,5 +184,6 @@ export const useCustomers = () => {
     clearFilter,
     sorting,
     filters,
+    customersData,
   };
 };
