@@ -1,13 +1,8 @@
-import { Nunito } from "next/font/google";
-
 import { LogoIcon } from "../../assets/icons/LogoIcon";
 import { useAppStore } from "../../store/appStore";
 import { useMediaQuery } from "../../hooks/useMediaQuery";
+import { nunito } from "../../styles/fonts";
 
-const nunito = Nunito({
-  weight: "700",
-  subsets: ["latin"],
-});
 export const Logo = () => {
   const { isSideMenuOpen, toggleSideMenu } = useAppStore();
   const isDesktop = useMediaQuery("(min-width: 1280px)");
@@ -15,7 +10,7 @@ export const Logo = () => {
   return (
     <div
       className={`h-20  
-        fill-mainColor dark:fill-mainColorDark text-center flex justify-center items-center text-[1.4rem] font-inter font-bold ${nunito.className}`}
+        fill-mainColor dark:fill-mainColorDark text-center flex justify-center items-center text-[1.4rem] font-bold ${nunito.className}`}
     >
       <LogoIcon />
       {(isSideMenuOpen || !isDesktop) && (
@@ -24,7 +19,7 @@ export const Logo = () => {
             Spire
           </div>
 
-          <div className="text-mainColor dark:text-mainColorDark  tracking-wider">
+          <div className="text-mainColor dark:text-mainColorDark tracking-wider">
             flow
           </div>
         </>
